backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-genai>=1.0.0
python-dotenv>=1.0.0
